:recycle: Split up file schema
