+++ creator/schema.py
@@ -10,7 +10,11 @@ from django_filters import FilterSet, OrderingFilter
 
 from creator.models import Job
 
-from creator.files import schema as file_mutations
+from creator.files.schema import (
+    FileMutation,
+    VersionMutation,
+    DownloadMutation,
+)
 from creator.studies.schema import Mutation as StudyMutation
 from creator.users import schema as user_mutations
 import creator.events.schema
@@ -243,7 +247,9 @@ class Status(graphene.ObjectType):
 
 
 class Query(
-    file_mutations.Query,
+    creator.files.schema.FileQuery,
+    creator.files.schema.VersionQuery,
+    creator.files.schema.DownloadQuery,
     creator.studies.schema.Query,
     creator.users.schema.Query,
     creator.events.schema.Query,
@@ -265,32 +271,13 @@ class Query(
         return Status(name="Kids First Study Creator", **info)
 
 
-class Mutation(StudyMutation, graphene.ObjectType):
-    create_file = file_mutations.file.FileUploadMutation.Field(
-        description="Upload a new file to a study"
-    )
-    create_version = file_mutations.version.VersionUploadMutation.Field(
-        description="Upload a new version of a file"
-    )
-    update_file = file_mutations.file.FileMutation.Field(
-        description="Update a file"
-    )
-    delete_file = file_mutations.file.DeleteFileMutation.Field(
-        description="Delete a file"
-    )
-    update_version = file_mutations.version.VersionMutation.Field(
-        description="Update a file version"
-    )
-    signed_url = file_mutations.SignedUrlMutation.Field(
-        description="Create a new signed url"
-    )
-    create_dev_token = file_mutations.DevDownloadTokenMutation.Field(
-        description="Create a new developer token"
-    )
-    delete_dev_token = file_mutations.DeleteDevDownloadTokenMutation.Field(
-        description="Delete a developer token"
-    )
-
+class Mutation(
+    StudyMutation,
+    FileMutation,
+    VersionMutation,
+    DownloadMutation,
+    graphene.ObjectType,
+):
     subscribe_to = user_mutations.SubscribeToMutation.Field(
         description="Subscribe the current user to a study"
     )
